+++ MarketPlace/src/main/java/com/challenges/MarketPlace/useCase/domain/Departamento.java
@@ -4,12 +4,14 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.util.List;
+
 @Getter
 @Setter
 @Builder
 public class Departamento {
 
-    private Long idDepartamento;
+    private Integer idDepartamento;
     private String nomeDepartamento;
     private String descricaoDepartamento;
 }
